extract common email exporter code
