+++ spec/workers/emails_users_export_worker_spec.rb
@@ -5,59 +5,6 @@ describe EmailsUsersExportWorker do
 
   it { is_expected.to be_a Sidekiq::Worker }
 
-  shared_examples "an email dump exporter" do
-
-    RSpec::Matchers.define :a_formatted_user_email do |x|
-      match do |csv_row|
-        csv_row.is_a?(Array) &&
-        csv_row.length == 1
-        csv_row.first.match(/.+@.+/)
-      end
-    end
-
-    let(:users) { create_list(:user, 2) }
-    let(:inactive_user) { create(:user, activated_state: :inactive) }
-    let(:invalid_email_user) { create(:user, valid_email: false) }
-
-    before do
-      users
-      inactive_user
-      invalid_email_user
-    end
-
-    after do
-      worker.perform(export_type)
-    end
-
-    it "should create a csv file with the correct number of entries" do
-      expect_any_instance_of(CSV)
-        .to receive(:<<)
-        .with(a_formatted_user_email)
-        .exactly(users.length)
-        .times
-    end
-
-    it "should compress the csv file" do
-      expect(worker).to receive(:to_gzip).and_call_original
-    end
-
-    it "push the file to s3 at the correct bucket location" do
-      adapter = MediaStorage::TestAdapter.new
-      allow(MediaStorage).to receive(:adapter).and_return(adapter)
-      expect(MediaStorage).to receive(:stored_path)
-        .with("application/x-gzip", "email_exports")
-        .and_call_original
-      expect(adapter)
-        .to receive(:put_file)
-        .with(s3_path, an_instance_of(String), s3_opts)
-      expect(worker).to receive(:write_to_s3).and_call_original
-    end
-
-    it "should clean up the csv and compressed files after sending to s3" do
-      expect(worker).to receive(:remove_tempfile).twice
-    end
-  end
-
   it_behaves_like "an email dump exporter" do
     let!(:non_global_user) { create(:user, global_email_communication: false) }
     let(:s3_path) { "email_exports/global_email_list.tar.gz" }
@@ -68,7 +15,7 @@ describe EmailsUsersExportWorker do
         content_disposition: "attachment; filename=\"global_email_list.csv\""
       }
     end
-    let(:export_type) { :global }
+    let(:export_params) { :global }
   end
 
   it_behaves_like "an email dump exporter" do
@@ -81,6 +28,6 @@ describe EmailsUsersExportWorker do
         content_disposition: "attachment; filename=\"beta_email_list.csv\""
       }
     end
-    let(:export_type) { :beta }
+    let(:export_params) { :beta }
   end
 end
